backend/requirements.txt
Flask
flask-cors
python-dotenv
supabase
requests
hdbscan
nltk
sentence_transformers
openai